Fixes cos, sin and tan calculations on angles
Fixes an issue with the cached cos, sin and tan calculation.  This used a piece of code that came from Vincenty and provided a very fast calculation.  Unfortunately, the code did not correctly provide the sign of the argument.  This does not matter for the Vincenty code, but did matter for other code.